Added support to use cg voices off the disk
Voice management has now improved. Linked-in voices are handled along
with clustergen voices that can be dynamically read off the disk.

When an application tries to set Language, the cg management part will
check that the voice file exists and load it if required.

THIS HAS NOT BEEN DEBUGGED YET.