Update instructions in Readme.rst for 2019
Instructions are very outdated and very difficult to build because (ndk r10r is no longer available and ant build.xml is no longer included in SDKs). I've updated instructions for 2019. I gave instructions on how to modify 10e so that it would work with your scripts and where to find missing build.xml for ant. You also have two different NDK (10d and 10e) and SDK (22 and 24.0.2) requirements in instructions. To avoid confusion I've made so only 10e and 24.0.2 are required. Please merge. Thanks!